[minor] Fix ResolutionMonitor and Add React SPA (#5)
While working with React SPA example, A several problems were
identified.

* ResolutionMonitor detects false postives, when resolution happens
concurrently.
* In React example, we need resolution cache for a promise so that
Suspense isn't stuck in an inifinite look when resolution fails.
* ResolutionError must be differentiated betweeen retriable error vs
something or something that needs to just fail.
* Boilerplate codes for maintaing list of all wires were very
cumbersome.
* Needing to have better best practices around wiring generic type to
concrete implementation, as we planned to not support tagged resolution
(such as using name, and etc)

This is great, because many of problems above were overlooked, or deemed
unnecessarily to tackle, which was clearly wrong.

In response to above problems, followings were added.

* ResolutionMonitor is going to be created on each resolution request,
thus we are free of concurrency concern.
* Top level TypeWireContainer will create ScopedResolutionContext in
which will have its own requestId and newly created ResolutionMonitor.
This ensures that a resolution request is tracked in its own request
scope, while tapping into the parent's state of all bindings, creators,
and loaded singleton instances if needed. (Hmm, we might have to
comeback and cache the resolver promise to prevent duplciated creation
of singleton instances.
* TypeWire now takes `imports` field with `createWith` method. This
greatly helps ergonomics of keeping track of dependent wires and not
needing to have full list of wires explicitly declared in everywhere.
This would deserve yet another best practice example for keeping track
of entire application specific wires though. We are currently doing it
through TypeWireGroup, but I am starting to think it may not be
necessarily, and we may have a better solution. I will follow up with
that in subsequent PRs.

Additionally, I believe TypeWireError needs to be differenciated between
Retryable vs not in subsequent PR, as well as the official react
integration.